add cc

+++ platform/consensus/recovery/recovery_test.cpp
@@ -1,20 +1,26 @@
 /*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
+ * Copyright (c) 2019-2022 ExpoLab, UC Davis
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without
+ * restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or
+ * sell copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following conditions:
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+ * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+ * DEALINGS IN THE SOFTWARE.
+ *
  */
 
 #include "platform/consensus/recovery/recovery.h"
@@ -106,7 +112,7 @@ TEST_F(RecoveryTest, ReadLog) {
 
     EXPECT_EQ(list.size(), expected_types.size());
 
-    for (size_t i = 0; i < expected_types.size(); ++i) {
+    for (int i = 0; i < expected_types.size(); ++i) {
       EXPECT_EQ(list[i].type(), expected_types[i]);
     }
   }
@@ -147,7 +153,7 @@ TEST_F(RecoveryTest, ReadLog_FlushOnce) {
 
     EXPECT_EQ(list.size(), expected_types.size());
 
-    for (size_t i = 0; i < expected_types.size(); ++i) {
+    for (int i = 0; i < expected_types.size(); ++i) {
       EXPECT_EQ(list[i].type(), expected_types[i]);
     }
   }
@@ -213,7 +219,7 @@ TEST_F(RecoveryTest, CheckPoint) {
 
     EXPECT_EQ(list.size(), types.size() * 14);
 
-    for (size_t i = 0; i < expected_types.size(); ++i) {
+    for (int i = 0; i < expected_types.size(); ++i) {
       EXPECT_EQ(list[i].type(), expected_types[i]);
     }
   }
@@ -290,7 +296,7 @@ TEST_F(RecoveryTest, CheckPoint2) {
 
     EXPECT_EQ(list.size(), types.size() * 14);
 
-    for (size_t i = 0; i < expected_types.size(); ++i) {
+    for (int i = 0; i < expected_types.size(); ++i) {
       EXPECT_EQ(list[i].type(), expected_types[i]);
     }
 
@@ -327,7 +333,7 @@ TEST_F(RecoveryTest, CheckPoint2) {
 
     EXPECT_EQ(list.size(), types.size() * 9);
 
-    for (size_t i = 0; i < expected_types.size(); ++i) {
+    for (int i = 0; i < expected_types.size(); ++i) {
       EXPECT_EQ(list[i].type(), expected_types[i]);
     }
     EXPECT_EQ(recovery.GetMinSeq(), 30);
@@ -409,7 +415,7 @@ TEST_F(RecoveryTest, SystemInfo) {
 
     EXPECT_EQ(list.size(), types.size() * 14);
 
-    for (size_t i = 0; i < expected_types.size(); ++i) {
+    for (int i = 0; i < expected_types.size(); ++i) {
       EXPECT_EQ(list[i].type(), expected_types[i]);
     }
 
@@ -449,7 +455,7 @@ TEST_F(RecoveryTest, SystemInfo) {
     EXPECT_EQ(data.primary_id(), 2);
     EXPECT_EQ(list.size(), types.size() * 9);
 
-    for (size_t i = 0; i < expected_types.size(); ++i) {
+    for (int i = 0; i < expected_types.size(); ++i) {
       EXPECT_EQ(list[i].type(), expected_types[i]);
     }
     EXPECT_EQ(recovery.GetMinSeq(), 30);
